# internal/api/assets/openapi.yaml
openapi: 3.0.3
info:
  title: Baseline API
  version: "1.0.0"
  description: "Baseline policy and dashboard API contract for implemented /v1 endpoints. Fields marked with x-baseline-implementation=reserved_noop are accepted for forward compatibility but not currently enforced."
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /v1/policies:
    get:
      summary: List policy summaries
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Policies
...
  /v1/rulesets:
    post:
      summary: Publish ruleset (admin)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRulesetRequest"
      responses:
        "201":
          description: Ruleset created
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/rulesets/{version}:
    get:
      summary: Get ruleset by version
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: version
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ruleset
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
    cookieAuth:
      type: apiKey
      in: cookie
      name: baseline_dashboard_session
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    CreateRulesetRequest:
      type: object
      required: [policy_names]
      properties:
        version:
          type: string
        description:
          type: string
        policy_names:
          type: array
          items:
            type: string
...